pyproject.toml
[project]
name = "openrouter_geminipro3_image_example"
version = "0.1.0"
description = "An example project using OpenRouter's Gemini Pro 3 image edditing."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=7.1.0",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
